+++ rest_framework_services/__init__.py
@@ -124,4 +124,4 @@ __all__ = [
     "update_from_input",
 ]
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"